Refactor SCRAM-SHA-1 unit tests: credential obfuscation tests
The fake salt determinism, fake key generation and user enumeration protection tests for SCRAM-SHA-1 are consolidated into one class, which is then
split up in an abstract base class, and a SCRAM-SHA-1 specific subclass. This is expected to make it easier to add test coverage for other algorithms.

Of note is that this change introduces a test-driven rather than domain-driven change in the system under test (`ScramSaslServer` and `ScramSha1SaslServer`): methods were lifted into the abstract superclass primarily because the tests needed polymorphism. This needs improving, which should go in a future commit.
